2798 3053 16486 수정

+++ rust/src/03/3053.rs
@@ -4,7 +4,7 @@ fn main() {
     let mut buf = String::new();
     std::io::stdin().read_line(&mut buf).unwrap();
 
-    let r: i32 = buf.trim().parse().unwrap();
+    let r: f64 = buf.trim().parse().unwrap();
 
-    println!("{:.6}\n{:.6}", (r * r as f64) * PI, (r * r as f64) * 2.0);
+    println!("{:.6}\n{:.6}", r * r * PI, r * r * 2.0);
 }
